Add some atrributes to crc32x64.v
The init_pipeline shift registers from multiple instances of the crc32x64
module have the potential to be merged/removed due to "equivalent register
removal", leading to large fanouts and subsequently poor timing.  This is not a
speculative claim; it has occurred in practice for the init_pipeline[4] bit.
The valid_pipeline has a similar potential, though that has not yet been seen
in practice.  Since these are only 7 FFs each, there is negligible harm in
preventing "equivalent register removal" for these shift registers, but lots of
potential benefits.

This adds the `(* equivalent_register_removal = "no" *)` attribute to the
init_pipeline and valid_pipeline shift registers in the crc32x64 module.  This
is the attribute for XST; other synthesis tools might need a differently named
attribute to accomplish the same purpose.