Resynth (again) fft_core netlist to limit fanout
The fft_core has some delays with write enables that go to BRAMs.  These write
enables are driven with a constant '1', but they pass through a "Delay" block
with a default (and unsettable) initial value of '0' so the constant '1' cannot
be absorbed into the BRAM.  Instead, the write enable Delays are all reduced to
a single register (due to equivalent register removal), which then fans out to
many (112) BRAMs.  This makes routing hard/impossible, so the fft_core netlist
was resynthesized using the "resynth_netlist" Matlab function from xps_library.

Resynthesizing with "-max_fanout 32" did not prevent the registers from being
reduced into one (the fanout was still 112), so this time we resynthesize with
"-equivalent_register_removal no".

The Matlab command line used was:

    resynth_netlist('fft_core', ...
      '-register_balancing', 'yes', ...
      '-optimize_primitives', 'yes' ...
      '-read_cores', 'yes' ...
      '-equivalent_register_removal', 'no'....
    )