First s6 BOF file (Maximum FPGA clock == 175 MHz)
This build failed timing, but creation of a bitstream was forced so we could
get the design running in the hardware.  This alloes for parallel integration
testing while the design/build is tweaked to get the clock rate up to the
desired rate of 225 MHz.

The worst case timing error was 1.215 ns longer than the requested 1.444 ns
period of the desired FPGA clock rate of 225 MHz.  This means the maximum FGPA
clock frequency for this design is 1000/(4.444+1.215) == 176.7 MHz, but let's
just call it 175 MHz, which corresponds to an ADC sample rate of 700 Msps.