Second s6 BOF file (Maximum FPGA clock == 195 MHz)
This build failed timing, but creation of a bitstream was forced so we could
get the design running in the hardware.  This allows for parallel integration
testing while the design/build is tweaked to get the clock rate up to the
desired rate of 225 MHz.

The minimum period reported by the timing tools is 5.066ns.  This means the
maximum FGPA clock frequency for this design is 1000/5.066 == 197.3 MHz, but
let's just call it 195 MHz, which corresponds to an ADC sample rate of 780
Msps.

This build has the fix for the "input/rms" BitBasher blocks.