Add auto and cross vzmac cores
A `vzmac` core is a vectorized-complex-multiply-accumulate core.  There are two
kinds of vzmac cores: an `auto vzmac` and a `cross vzmac`.  An `auto vzmac`
computes the vector sum of the power of each element of a complex input vector.
A `cross vzmac` computes the vector sum of the cross power of corresponding
elements of two input vectors.

These vzmac cores are configured for serendip6/alfaburst, so they have a vector
length of 2048 and they sum 14 vectors per accumulation.  They have a built-in
controller to drive this timing, so all they need is a single sync pulse on the
cycle before the first sample of the first vector.

The vzmac cores use DSP48 slices to their full potential.  The auto vzmac
operation uses 2 DSP48 slices and the cross vzmac operation uses 4 DSP48
slices.  Each of these DSP48 slices uses a multiply/add (or multiply/subtract)
operation, so no fabric adders are required at all for the vzmac operation.
The counter in the delay_blocks also use a DSP48 slice, so the total DSP48
usage per core is 3 for the auto and 6 for the cross.

This also adds test scripts to verify the output of the simulated core models.