Consolidate crc32x64 implementations
The inital copy from the PAPER F engine design included the `crc32x64`
subdirectory.  When the `s6_apcketizer` code was added later, it also
included a copy of the `crc32x64` funcitonality in the `verilog`
subdirectory and ignored the code in the `crc32x64` subdirectory.  This
was not a problem until the `ab_packetizer` logic was added to `s6.mdl`
and referenced the files in the `crc32x64` subdirectory.  That led to
the inclusion of two equivalent but distinct `crc32x64` modules which
led to a redefinition error during synthesis of the system generator
model.

This removes the `crc32x64` subdirectory and changes the
`crc32x64_config.m` file to use the code from the `verilog`
subdirectory.