Tighten up the pkt_pblock
Using a whole clock region covers too wide of an area.  I think the
S6 packetizer is not optimal in terms of pipelining.  I also think part
of the problem is that parts from one of the packatizer instances get
used for all packetizers (maybe thanks to equivalent register removal)
so that leads to fan out issues.  The packetizer can run at the desired
speed so long as it is placed compactly enough.  That's what this change
to the pkt_pblock attempts to do.