Add more equivalent_register_removal to verilog
The `s6_packetizer.v` file has an `equivalent_register_removal`
attribute preceding the module declaration, but I think that might not
sufficient.  It is apparently being recognized during synthesis by XST
because the `s6/sysgen/synth_model/s6.results` file contains this text:

    Synthesizing Unit <s6_packetizer>.
        Related source file is "/home/davidm/casper/s6_gateware/s6/sysgen/synth_model/s6_packetizer.v".
            NWORDS = 2048
        Set property "equivalent_register_removal = no".

I think that it might only apply to the intances as a whole, not to the
individual registers within the instances, because the same log file contains
lots of text like:

    INFO:Xst:2261 - The FF/Latch
    <packetize_bc30b44243/packetizer0_23a7528456/packtizer/word_count_0>
    in Unit <s6_x0> is equivalent to the following 3 FFs/Latches, which
    will be removed :
    <packetize_bc30b44243/packetizer1_7f020d3e98/packtizer/word_count_0>
    <packetize_bc30b44243/packetizer2_fadff61cae/packtizer/word_count_0>
    <packetize_bc30b44243/packetizer3_e09b6d4efc/packtizer/word_count_0>

This commit adds the equivalent_register_removal attribute explicitly to
the registers in the `s6_packetizer` module that were listed in the log
file as being removed due to equivalency.