Prevent "data bubble" in AB packetizer FIFOs
The AB packetizer requires that only 2048 words are written to its FIFOs
every integration.  It turns out that the first integration at startup
outputs 2048+X (X>0) valid samples due to how the vector accumulators
work.  These X extra samples form a sort of "data bubble" in the FIFOs.
After each integration, 2048 words are read out of the FIFOs.  These
include words left over from from the previous integration (the "bubble
words) and then 2048-X words of the current integration, but X words
from the current integration are left behind as the next bubble.

This change adds a "valid gate" before the packetizer's data_valid input
to ensure that the data_valid signal is always exactly 2048 cycles long,
even if the vector accumulators' data valid output is not exactly 2048
cycles long.  This prevents the formation of a data bubble in the AB
packetizer FIFOs.