Add power/snap snapshot block
This will allow one to take snapshots of the even channels of cross
vzmac output for beam 1 XY* (i.e. inputs 2 and 3).  Using the external
trigger and write enable will capture only "valid" outputs starting
after a falling edge of "valid", so data should be aligned with the
start of a spectrum (i.e. the "always real" channel 0 should be the
first channel captured if everything is working properly).