Add FIFO status monitoring to ab_packetizer_test
This adds diagnostic monitoring of the FIFO status to the
ab_packetizer_test.mdl file.  This model produces the desired/expected
results when simulating, but "identical" logic (or at least "should be
identical" logic) in the s6.mdl file does not give the same results.

The primary symptom observed when running the AB packetizer in the S6
model is that the DC channel (always zero imaginary component in XY*)
appears to be in channel 4 rather than channel 0.